Extract projectsPicker
